New Class : Weapon
Weapon Class
