+++ Assets/Scenes/TestWeapon.unity.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dea18a255ae70ec4683944b72c12d3a0
+timeCreated: 1485393236
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
