+++ Assets/Scripts/Weapons/DualBlastersScript.cs
@@ -2,23 +2,8 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-public class DualBlastersScript : MonoBehaviour {
+public class DualBlastersScript : Weapon {
 
-    public string name = "DualBlasters";
-    public int range;
-    public int damage;
-    public float fireRate;
-    public int ammo;
-
-    public GameObject projectile;
-    public float shotSpeed;
-    public Transform spawnPoint;
-    public GameObject Reticle;
-    public GameObject Gauge;
-    
-    public GameObject Particleprefab;
-
-    private float lastShot;
 
     // Use this for initialization
     void Start () {
@@ -27,28 +12,6 @@ public class DualBlastersScript : MonoBehaviour {
 	// Update is called once per frame
 	void Update () {
 
-    }
-
-    public void OnFire()
-    {
-        if (lastShot + fireRate < Time.time)
-        {
-            GameObject clone;
-            clone = Instantiate(projectile, spawnPoint.position, spawnPoint.rotation);
-            clone.GetComponent<ProjectileScript>().SetParent(Reticle,Gauge,this.gameObject);
-            clone.GetComponent<ProjectileScript>().damage = damage;
-
-            clone.GetComponent<Rigidbody>().velocity = spawnPoint.forward * shotSpeed;
-            clone.GetComponent<Rigidbody>().velocity += spawnPoint.GetComponentInParent<Rigidbody>().velocity;
-            lastShot = Time.time;
-        }
-
-    }
-
-    public void DestroyEnemyAnimation(Transform DeathTransform)
-    {
-        GameObject Particle = Instantiate(Particleprefab, DeathTransform.position, Random.rotation);
-        Destroy(Particle, 2);
     }
 
 }
