+++ Assets/Scripts/Weapons/Weapon.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: df845e373ff40d5458f066170ea62692
+timeCreated: 1490729653
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
